backend/requirements.txt
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
django-cors-headers==4.6.0
django-filter==24.3
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-redis==5.4.0
django-storages==1.14.4
django-import-export==4.3.3

celery[redis]==5.4.0
redis==5.2.1

psycopg2-binary==2.9.10
gunicorn==23.0.0

Pillow==11.0.0
boto3==1.36.2
python-decouple==3.8
requests==2.32.3
beautifulsoup4==4.12.3
premailer==3.10.0
html2text==2024.2.26

sentry-sdk==2.19.2
whitenoise==6.8.2
